Completed merchant as well as admin dashboard functionality, bug fixes
