+++ package.json
@@ -12,6 +12,7 @@
     "flexboxgrid": "^6.3.1",
     "material-ui-pickers": "^1.0.0-rc.14",
     "moment": "^2.22.2",
+    "public-ip": "^2.4.0",
     "randomcolor": "^0.5.3",
     "react": "^16.4.1",
     "react-chartjs-2": "^2.7.4",
